bootloader/requirements.txt
Celery
Django
django-fsm
django-netfields
djangorestframework
netaddr
rest-framework-generic-relations
psycopg2
pycrypto
requests
PyYAML
